backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
<<<<<<< HEAD
mediapipe==0.10.13
numpy==1.26.4
tensorflow==2.16.1
torch==2.2.2
torchvision==0.17.2
pillow==10.1.0
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.0
deepface
fer
tf-keras==2.16.0
setuptools<81
=======
mediapipe==0.10.9
numpy==1.24.3
tensorflow==2.14.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
scikit-learn==1.3.2
python-dotenv==1.0.0
cors==1.0.1
pydantic==2.5.0
>>>>>>> origin/brayden-frontend
